Make `format_timedelta` a helper function
